requirements.txt
selenium
django
requests
django-background-tasks